+++ module/userProjects/helper.js
@@ -25,6 +25,8 @@ const userProfileService = require(GENERICS_FILES_PATH + "/services/users");
 const solutionsHelper = require(MODULES_BASE_PATH + "/solutions/helper");
 const certificateTemplateQueries = require(DB_QUERY_BASE_PATH + "/certificateTemplates");
 const certificateService = require(GENERICS_FILES_PATH + "/services/certificate");
+const certificateValidationsHelper = require(MODULES_BASE_PATH + "/certificateValidations/helper");
+const _ = require("lodash");  
 
 /**
     * UserProjectsHelper
@@ -371,15 +373,9 @@ module.exports = class UserProjectsHelper {
                         status: HTTP_STATUS_CODE['bad_request'].status
                     }
                 }
-
-                // push to kafka only if project is submitted and certificate key is present
-                if ( projectUpdated.status == CONSTANTS.common.SUBMITTED_STATUS &&
-                     projectUpdated.certificate &&
-                     Object.keys(projectUpdated.certificate).length > 0
-                ) {
-                    await kafkaProducersHelper.pushProjectToKafka(projectUpdated);
-                }
-                
+                //  push project details to kafka
+                await kafkaProducersHelper.pushProjectToKafka(projectUpdated);
+            
                 return resolve({
                     success: true,
                     message: CONSTANTS.apiResponses.USER_PROJECT_UPDATED,
@@ -1172,12 +1168,7 @@ module.exports = class UserProjectsHelper {
                         
                         // create certificate object and add data if certificate template is present.
                         if ( certificateTemplateDetails.length > 0 ) {
-                            projectCreation.data["certificate"] = {
-                                templateId : certificateTemplateDetails[0]._id,
-                                templateUrl : certificateTemplateDetails[0].templateUrl,
-                                status : certificateTemplateDetails[0].status,
-                                criteria : certificateTemplateDetails[0].criteria
-                            }
+                            projectCreation.data["certificate"] = _.pick(certificateTemplateDetails[0], ['_id', 'templateUrl', 'status', 'criteria']);
                         }
                     }
                     
@@ -2223,12 +2214,7 @@ module.exports = class UserProjectsHelper {
                     
                     // create certificate object and add data if certificate template is present.
                     if ( certificateTemplateDetails.length > 0 ) {
-                        libraryProjects.data["certificate"] = {
-                            templateId : certificateTemplateDetails[0]._id,
-                            templateUrl : certificateTemplateDetails[0].templateUrl,
-                            status : certificateTemplateDetails[0].status,
-                            criteria : certificateTemplateDetails[0].criteria,
-                        }
+                        libraryProjects.data["certificate"] = _.pick(certificateTemplateDetails[0], ['_id', 'templateUrl', 'status', 'criteria']);
                     }
                     delete  libraryProjects.data.solutionInformation.certificateTemplateId;
                 }
@@ -2340,54 +2326,6 @@ module.exports = class UserProjectsHelper {
     }
 
     /**
-     * validate certificate criteria.
-     * @method
-     * @name criteriaValidation 
-     * @param {Object} data - project data for certificate creation 
-     * @returns 
-    */
-
-    static criteriaValidation(data) {
-        return new Promise(async (resolve, reject) => {
-            try {
-                let criteria = data.certificate.criteria;
-                let validationResult = [];
-                let validationMessage = "";
-                let validationExpression = criteria.expression
-                if ( criteria.conditions &&  Object.keys(criteria.conditions).length > 0 ) {
-                    let conditions = criteria.conditions;
-                    let conditionKeys = Object.keys(conditions)
-
-                    for ( let index = 0; index < conditionKeys.length; index++ ) {
-                        // correntCondition contain the prefinal level data
-                        let currentCondition = conditions[conditionKeys[index]];
-
-                        //now pass expression and validation scope to another function which will start the validation procedure
-                        let validation = await _subCriteriaValidation( currentCondition.conditions, currentCondition.expression, data );
-                        
-                        validationResult.push(validation.success);
-                        ( validation.success == false ) ? validationMessage = validationMessage + " " + currentCondition.validationText : "";
-                    }
-                    let criteriaValidation = await _criteriaExpressionValidation( validationExpression, conditionKeys, validationResult )
-                    return resolve({
-                        success: criteriaValidation,
-                        message: ( criteriaValidation == false ) ? validationMessage : CONSTANTS.common.PROJECT_CERTIFICATE_GENERATED_SUCCESSFULLY
-                    });
-                }
-                return resolve({
-                    success: false
-                })
-            } catch (error) {
-                return resolve({
-                    success: false,
-                    message: error.message,
-                    data: {}
-                });
-            }
-        })
-    }
-
-    /**
      * generate project certificate.
      * @method
      * @name generateCertificate 
@@ -2400,7 +2338,7 @@ module.exports = class UserProjectsHelper {
             try {
                 //  if eligible key is not there check criteria for validation
                 if ( !data.certificate.eligible ) {
-                    let validateCriteria = await this.criteriaValidation(data)
+                    let validateCriteria = await certificateValidationsHelper.criteriaValidation(data)
                     if ( validateCriteria.success ) {
                         data.certificate.eligible = true;
                     } else {
@@ -2479,6 +2417,7 @@ module.exports = class UserProjectsHelper {
                         };
                         
                         const certificateDetails = await certificateService.createCertificate( certificateData );
+                        
                         if ( !certificateDetails.success || !certificateDetails.data || !certificateDetails.data.ProjectCertificate ) {
                             return resolve({
                                 success:false
@@ -2734,163 +2673,6 @@ module.exports = class UserProjectsHelper {
 
 
 };
-/**
- * _subCriteriaValidation.
- * @method
- * @name _subCriteriaValidation 
- * @param {Object} conditions - condition data.
- * @param {String} expression - validation expression
- * @returns {Boolean} validation.
-*/
-
-function _subCriteriaValidation(conditions, expression, data) {
-   return new Promise(async (resolve, reject) => {
-        try {
-            let conditionKeys = Object.keys(conditions)
-            let validationResult = [];
-           
-            for ( let index = 0; index < conditionKeys.length; index++ ) {
-                let currentCondition = conditions[conditionKeys[index]];
-                // correntCondition contain the prefinal level data
-                //now pass expression and validation scope to another function which will start the validation procedure
-                let validation = await _validateCriteriaConditions( currentCondition, data );
-                validationResult.push(validation);
-            }
-            
-            let subcriteriaValidation = await _criteriaExpressionValidation( expression, conditionKeys, validationResult )
-            return resolve({
-                success: subcriteriaValidation
-            });
-
-        } catch (error) {
-            return resolve({
-                message: error.message,
-                success: false,
-                status:
-                    error.status ?
-                        error.status : HTTP_STATUS_CODE['internal_server_error'].status
-            })
-        }
-    })
-}
-
-/**
- * _validateCriteriaConditions.
- * @method
- * @name _validateCriteriaConditions 
- * @param {Object} condition - condition data.
- * @param {String} data - validation data
- * @returns {Boolean} validation.
-*/
-
-function _validateCriteriaConditions(condition, data) {
-    return new Promise(async (resolve, reject) => {
-        try {
-            let result = false;
-            if ( !condition.function || condition.function == "" ) {
-                if ( condition.scope == CONSTANTS.common.PROJECT ){
-
-                    // let expression = data[condition.key] + condition.operator + condition.value;
-                    if ( condition.key == "completedDate") {
-                        let comparableDates = UTILS.createComparableDates( data[condition.key], condition.value );
-                        data[condition.key] = comparableDates.dateOne;
-                        condition.value = comparableDates.dateTwo;
-                    }
-                    result = UTILS.operatorValidation( data[condition.key], condition.value, condition.operator );
-                    
-                } 
-            } else {
-                try {
-                    let valueFromProject = 0;
-                    // if: condition is in scope of project and contains a function to check
-                    if ( condition.scope == CONSTANTS.common.PROJECT ) {
-                        valueFromProject = UTILS.getAttachmentCount( data[condition.key], condition.filter ); 
-                    } else if ( condition.scope == CONSTANTS.common.TASK_ATTACHMENT ){
-                        // for task attachment validatiion _id of specific task or "all" key should be passed in an array called taskDetails
-                        let tasksAttachments = [];
-                        let projectTasks = data.tasks;
-                        
-                        if ( projectTasks && projectTasks.length > 0 && condition.taskDetails.length > 0 &&  condition.taskDetails[0] == "all" ) {
-                            
-                            for ( let tasksIndex = 0; tasksIndex < projectTasks.length; tasksIndex++ ) {
-
-                                if ( projectTasks[tasksIndex][condition.key] && projectTasks[tasksIndex][condition.key].length > 0 ) 
-                                {
-                                    tasksAttachments.push(...projectTasks[tasksIndex][condition.key])
-                                }
-                            }
-
-                        } else if ( projectTasks && projectTasks.length > 0 && condition.taskDetails.length > 0 ) {
-
-                            // specific task Id or Ids are passed for attachment validation
-                            for ( let tasksIndex = 0; tasksIndex < projectTasks.length; tasksIndex++  ) {
-                                for ( let taskDetailsPointer = 0; taskDetailsPointer < condition.taskDetails.length; taskDetailsPointer++ ) {
-                                    // get attachments data of specified task/ tasks
-                                    if ( projectTasks[tasksIndex]._id == condition.taskDetails[taskDetailsPointer] && projectTasks[tasksIndex][condition.key] && projectTasks[tasksIndex][condition.key].length > 0 ) {
-                                        tasksAttachments.push(...projectTasks[tasksIndex][condition.key])
-                                    }
-                                }
-                                
-                            }
-
-                        } else {
-                            return resolve(result)
-                        }
-                        if ( !tasksAttachments.length > 0 ) {
-                            return resolve(result)
-                        }
-                        valueFromProject = UTILS.getAttachmentCount( tasksAttachments, condition.filter ); 
-                    }
-                    result =  UTILS.operatorValidation( valueFromProject, condition.value, condition.operator );
-
-                } catch (fnError) {
-                    return resolve(result)
-                }
-            }            
-            return resolve(result);
-        } catch (error) {
-            return resolve({
-                message: error.message,
-                success: false,
-                status:
-                    error.status ?
-                        error.status : HTTP_STATUS_CODE['internal_server_error'].status
-            })
-        }
-    })
-}
-/**
- * _criteriaExpressionValidation
- * @method
- * @name _criteriaExpressionValidation 
- * @param {String} expression - criteria expression
- * @param {Array} keys - condition keys
- * @param {Array} result - condition result
- * @returns {Boolean} validation result.
-*/
-
-function _criteriaExpressionValidation(expression, keys, result) {
-    return new Promise(async (resolve, reject) => {
-        try {
-            
-            if( expression == "" ||
-                !keys.length > 0 ||
-                !result.length > 0 ||
-                keys.length != result.length ) {
-                return resolve(false);
-            }
-            for ( let pointerToKeys = 0; pointerToKeys < keys.length; pointerToKeys++ ) {
-                expression = expression.replace(keys[pointerToKeys],result[pointerToKeys].toString())
-            }
-            let evalResult = eval(expression)
-            
-            return resolve(evalResult);
-
-        } catch (error) {
-            return resolve(false);
-        }
-    })
-}
 
 /**
  * Project information.
